feat: add useFetch custom hook with dynamic fetch url
