+++ simple-tab/src/practices/Fetch/useFetch.js
@@ -1,0 +1,29 @@
+import { useState, useEffect } from "react";
+
+export default function useFetch(url) {
+    const [data, setData] = useState(null);
+    const [loading, setLoading] = useState(true);
+    const [error, setError] = useState(null);
+
+    useEffect(() => {
+        const fetchData = async () => {
+            setLoading(true);
+            try {
+                const response = await fetch(url);
+                const result = await response.json();
+                setData(result);
+            } catch (err) {
+                setError(err);
+            } finally {
+                setLoading(false);
+            }
+        };
+        fetchData();
+    }, [url]);
+
+    return {
+        data,
+        loading,
+        error,
+    }
+}
